api:file: add API for file-system file I/O
The file I/O API provides the means to perform stateless, that is,
positional read/write, to file-system files. A file-specific
open-function is provided to give the user control over whether or not
operating system caches are uses.
This translates into backend handling files with O_DIRECT on Linux, when
enabled, and similar OS-specific mechanism for enabling/disabling
Operating System page-cache for file-operations. To facilitate
operating-system cache-flush then a fsync primitive is provided as well.

The API utilizes the regular xNVMe synchronous and asynchronous command
interface.

Signed-off-by: Michael Bang <mi.bang@samsung.com>
Signed-off-by: Simon A. F. Lund <simon.lund@samsung.com>