file: initial implementation of api:file
The file API works with OS-managed files in file-systems. These files,
unlike block-devices, are byte-addressable, although page or
block-alignment can be required for some instances, then commonly it is
possible to open up a file are read/write it a byte at a time, this
collides with the block/LBA addressing assumed by the library backends
for the NVM command-set.

So, although the pread/pwrite operations are represented as NVM
read/write, then they do not translate to LBA addresses but instead
assign byte-offset to slba. It is thus up to the backend to handle this
accordingly.

The default-mode for xnvme_file_open(), is to open files without
direct=1.

Signed-off-by: Michael Bang <mi.bang@samsung.com>
Signed-off-by: Simon A. F. Lund <simon.lund@samsung.com>