feat(macos/driverkit): implement new DriverKit-based backend using MacVFN
This backend only works on macOS and depends on MacVFN being
installed on the system. MacVFN uses Apple's DriverKit to hook a
PCI device and by support of libvfn's code implements an NVMe driver.

MacVFN devices are currently enumerated by a counter in xNVMe, as
bus-device-function is seemingly not available.