fix(examples): simplify single IO examples
These examples will be used in the documentation, and should be as
simple as possible.

Thus, remove unnecessary printing of the payload for both examples.
And instead add printing of completed steps: submission and completions.

For the async example: remove pointless goto and cb_args struct.
Also simplify naming of some elements.

Signed-off-by: Karl Bonde Torp <k.torp@samsung.com>